+++ spec/views/projects/index.html.erb_spec.rb
@@ -1,0 +1,19 @@
+require 'rails_helper'
+
+RSpec.describe 'projects/index', type: :view do
+  before(:each) do
+    assign(:projects, [
+             Project.create!(
+               name: 'Name'
+             ),
+             Project.create!(
+               name: 'Name'
+             )
+           ])
+  end
+
+  it 'renders a list of projects' do
+    render
+    assert_select 'tr>td', text: 'Name'.to_s, count: 2
+  end
+end