+++ spec/views/projects/show.html.erb_spec.rb
@@ -1,0 +1,14 @@
+require 'rails_helper'
+
+RSpec.describe 'projects/show', type: :view do
+  before(:each) do
+    @project = assign(:project, Project.create!(
+                                  name: 'Name'
+                                ))
+  end
+
+  it 'renders attributes in <p>' do
+    render
+    expect(rendered).to match(/Name/)
+  end
+end