+++ Changelog.md
@@ -7,6 +7,7 @@
 * Ability to send environment variables to docker run command
 * Support any number of projects to run tests
 * Do not stop change branches if any branch non exists
+* Store `Projects` as model, not as config parameter 
 
 ### Fixes
 * Fix silent fail if timeout while restoring image